requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pandas
scikit-learn==1.3.2
joblib
kafka-python